requirements.txt
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==10.4.0
streamlit==1.36.0
tensorflow==2.13.1
